+++ apps/web-giddh/src/app/new-inventory/component/variant-wise/variant-wise.component.scss
@@ -1,4 +1,4 @@
-.varient-wise-wrapper {
+.variant-wise-wrapper {
     margin: 20px 20px 10px;
     .select-dropdown-wrapper {
         .select-dropdown {
