+++ apps/web-giddh/src/assets/css/style-2.scss
@@ -7108,9 +7108,9 @@ span#HW_badge_cont:nth-child(3) {
 }
 
 .item-wise-page,
-.varient-wise-page {
+.variant-wise-page {
     .item-wise-wrapper,
-    .varient-wise-wrapper {
+    .variant-wise-wrapper {
         .select-dropdown-wrapper {
             .select-field {
                 input {
